+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/plants/tree/leaves/BWGFruitLeavesBlock.java
@@ -14,12 +14,12 @@ import org.jetbrains.annotations.NotNull;
 import java.util.function.Supplier;
 
 public class BWGFruitLeavesBlock extends LeavesBlock implements BonemealableBlock {
-    private final Supplier<BlockState> fruitBlock;
+    private final Supplier<BWGFruitBlock> fruitBlock;
     private final float tickSpawnChance;
 
     public BWGFruitLeavesBlock(Properties properties, Supplier<BWGFruitBlock> fruitBlock, float tickSpawnChance) {
         super(properties);
-        this.fruitBlock =  () -> fruitBlock.get().defaultBlockState().setValue(BWGFruitBlock.AGE, 0);
+        this.fruitBlock = fruitBlock;
         this.tickSpawnChance = tickSpawnChance;
     }
 
@@ -29,34 +29,33 @@ public class BWGFruitLeavesBlock extends LeavesBlock implements BonemealableBloc
         BlockPos fruitPos = pos.below();
         if (this.decaying(state)) {
             BlockState below = level.getBlockState(fruitPos);
-            if (below.is(fruitBlock.get().getBlock()))
+            if (below.is(fruitBlock.get()))
                 level.destroyBlock(fruitPos, below.getValue(BWGFruitBlock.AGE) == BWGFruitBlock.MAX_AGE || random.nextBoolean());
         } else if (level.getBlockState(fruitPos).isAir() && random.nextFloat() < this.tickSpawnChance)
             placeFruit(level, fruitPos);
     }
 
-    private void placeFruit(Level level, BlockPos pos) {
-        level.setBlock(pos, fruitBlock.get(), 2);
+    private void placeFruit(@NotNull Level level, @NotNull BlockPos pos) {
+        level.setBlock(pos, fruitBlock.get().defaultBlockState().setValue(BWGFruitBlock.AGE, 0), 2);
     }
 
     @Override
     public boolean isValidBonemealTarget(@NotNull LevelReader level, @NotNull BlockPos pos, @NotNull BlockState state, boolean isClient) {
-        return true;
+        return !state.getValue(PERSISTENT);
     }
 
     @Override
-    public boolean isBonemealSuccess(Level level, @NotNull RandomSource random, BlockPos pos, @NotNull BlockState state) {
+    public boolean isBonemealSuccess(@NotNull Level level, @NotNull RandomSource random, @NotNull BlockPos pos, @NotNull BlockState state) {
         BlockState below = level.getBlockState(pos.below());
-            if (below.isAir() || below.getBlock() instanceof BWGFruitBlock && below.getValue(BWGFruitBlock.AGE) < 3)
+            if (below.isAir() || below.is(fruitBlock.get()) && below.getValue(BWGFruitBlock.AGE) < 3)
                 return random.nextBoolean();
         return false;
     }
 
     @Override
-    public void performBonemeal(ServerLevel level, @NotNull RandomSource random, BlockPos pos, @NotNull BlockState state) {
+    public void performBonemeal(@NotNull ServerLevel level, @NotNull RandomSource random, BlockPos pos, @NotNull BlockState state) {
         BlockState below = level.getBlockState(pos.below());
         if (below.isAir()) placeFruit(level, pos.below());
-        else if (below.is(fruitBlock.get().getBlock()))
-            level.setBlock(pos.below(), below.cycle(BWGFruitBlock.AGE), 2);
+        else if (below.is(fruitBlock.get())) level.setBlock(pos.below(), below.cycle(BWGFruitBlock.AGE), 2);
     }
 }